+++ src/ocm/components/clusterDetail/ClusterDetail.tsx
@@ -11,7 +11,6 @@ import {
   Alert,
 } from '@patternfly/react-core';
 import {
-  Cluster,
   ToolbarButton,
   ToolbarSecondaryGroup,
   Alerts,
@@ -20,7 +19,10 @@ import {
   RenderIf,
   KubeconfigDownload,
   REDHAT_CONSOLE_OPENSHIFT,
+  getEnabledHosts,
+  getOlmOperators,
 } from '../../../common';
+import { Cluster } from '../../../common/api/types';
 import ClusterHostsTable from '../hosts/ClusterHostsTable';
 import ClusterToolbar from '../clusters/ClusterToolbar';
 import { downloadClusterInstallationLogs, getClusterDetailId } from './utils';
@@ -40,6 +42,7 @@ import { TIME_ZERO, VSPHERE_CONFIG_LINK } from '../../../common/config/constants
 import { isSNO } from '../../../common/selectors/clusterSelectors';
 import { ExternalLinkAltIcon } from '@patternfly/react-icons';
 import { diffInDaysBetweenDates } from '../../../common/sevices/DateAndTime';
+import { getClusterProgressAlerts } from './getProgressBarAlerts';
 
 type ClusterDetailProps = {
   cluster: Cluster;
@@ -61,6 +64,7 @@ const ClusterDetail: React.FC<ClusterDetailProps> = ({ cluster }) => {
   const { inactiveDeletionHours } = useDefaultConfiguration(['inactiveDeletionHours']);
   const inactiveDeletionDays = Math.round((inactiveDeletionHours || 0) / 24);
   const dateDifference = calculateDateDifference(inactiveDeletionDays, cluster.installCompletedAt);
+  const { monitoredOperators = [] } = cluster;
 
   return (
     <Stack hasGutter>
@@ -73,11 +77,18 @@ const ClusterDetail: React.FC<ClusterDetailProps> = ({ cluster }) => {
           </GridItem>
           <GridItem span={7}>
             <ClusterProgress
-              consoleUrl={credentials?.consoleUrl}
               cluster={cluster}
               onFetchEvents={onFetchEvents}
               totalPercentage={cluster.progress?.totalPercentage || 0}
             />
+          </GridItem>
+          <GridItem span={7}>
+            {getClusterProgressAlerts(
+              getEnabledHosts(cluster.hosts),
+              cluster,
+              getOlmOperators(monitoredOperators),
+              credentials?.consoleUrl,
+            )}
           </GridItem>
           <ClusterDetailStatusVarieties cluster={cluster} clusterVarieties={clusterVarieties} />
           <RenderIf condition={dateDifference > 0}>
